+++ core-configuration/identities/identityAssignment.sqf
