+++ startup/configuration/internals/configMacros.hpp
@@ -10,7 +10,7 @@
 #include "..\..\..\configuration\killTracking.hpp"
 #include "..\..\..\configuration\gravestones.hpp"
 #include "..\..\..\configuration\viewDistanceEditor.hpp"
-#include "..\..\..\configuration\identityReplacement.hpp"
+#include "..\..\..\core-configuration\identityReplacement.hpp"
 #include "..\..\..\configuration\insignia.hpp"
 #include "..\..\..\configuration\objectScaling.hpp"
 #include "..\..\..\configuration\sogConfig.hpp"
