+++ startup/configuration/internals/identityTranslation.sqf
@@ -5,8 +5,8 @@ if (isNil "f_var_identity_already_run") then
     #include "..\..\..\identity_macros.hpp"
     f_var_identity_already_run = true;
 
-    #include "..\..\..\configuration\identities\identityGroups.sqf"
-    #include "..\..\..\configuration\identities\identityAssignment.sqf"
+    #include "..\..\..\core-configuration\identities\identityGroups.sqf"
+    #include "..\..\..\core-configuration\identities\identityAssignment.sqf"
 
     f_var_identity_loaded = true;
 }
